# Enhanced OCI Registry Authentication and Layer Extraction Support